Verify Hashicorp GPG signatures and checksums
This plugin did not check GPG signatures or checksums. So when the
CodeCov breach affected Hashicorp, leading to

https://discuss.hashicorp.com/t/hcsec-2021-12-codecov-security-event-and-hashicorp-gpg-key-exposure/23512

and a GPG key update announcement on

https://www.hashicorp.com/security

the first thing I checked was whether my asdf plugin for Hashicorp Vault
had been updated to support the new key.

Alas, it hadn't! In fact, asdf wasn't verifying GPG signatures at all,
meaning that supply chain compromises of Hashicorp infrastructure could
trivially deliver compromised Vault, Terraform, and Consul binaries to
users.

To fix this, I've added signature and checksum verification to
asdf-hashicorp. The GPG key is static, included in this repo, because in
modeling a future attack where the adversary has control of HashiCorp's
webservers (but not their GPG notary server), we would remain at risk if
we simply fetched the latest key each time. Instead, maintainers of this
repository will need to update the GPG key in this repo when it expires
or is revoked because of a security incident.