conditionally verify different signature files
I believe this to be a robust method of signature file selection that
works for all existing signature file paths, including the new case for
today (where the keyid suffix is not present at all).

In addition, add a test suite for installing the current first and last
Terraform/Consul/Vault binaries.

    $ make test
    bats test
    ✓ install command fails if the input is not version number
    ✓ can install and verify earliest terraform
    ✓ can install and verify latest terraform
    ✓ can install and verify earliest vault
    ✓ can install and verify latest vault
    ✓ can install and verify earliest consul
    ✓ can install and verify latest consul

    7 tests, 0 failures