Guide user if error on finding the version
Add a repository link to the error message so the user can follow and better figure out how to workaround. That comes with the confusion on missing darwin arm64 builds for old versions of terraform. There is already a variable to bypass it and run on compatibility mode. Point a link can avoid extra support on chats or other channels.